requirements.txt
# Core dependencies
numpy==1.21.0
matplotlib==3.4.2

# Text processing
nltk==3.6.2
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.3.4
beautifulsoup4==4.9.3

# URL handling
requests==2.25.1

# LaTeX generation
jinja2==3.0.1

# Development dependencies
pylint==2.8.3
black==21.6b0